sapper bug(?): need to use underscore with slug
for some reason, when using dashes in songs-arent-time-travel slug,
the songs-arent-time-travel.json isn't created in sapper
export.. which leads to a very confusing bug where navigating to the
piece from index fails with something like "json failed, unexpected <
character". *but*, when i navigate directly to that page, or refresh
the page, the load works. (i think this has to do with SPA-ish
navigation on links, as the actual index.html file _is_ built).

   5.26 kB   pieces/songs_arent_time_travel/index.html
   23.8 kB   pieces/arturo/index.html
   24.4 kB   pieces/arturo.json
   # when using dash slug, this file isn't created
   1.93 kB   pieces/songs_arent_time_travel.json