dont use concurrent requests on sapper export!!
i've spent so much time trying to figure out why sapper export often
misses exporting the [slug].json for certain pages. in b0b3b83 i
assumed the issue was (somehow) using hyphens in slugs so switched
to using underscores but this started happening again.

i found the followng issues which all are "closed" or "merged" but
this is _definitely_ still a bug. ended up looking at options for the
sapper export command and saw i can turn off concurrent requests,
which seems to have fixed the issue.
- https://github.com/sveltejs/sapper/issues/1078
- https://github.com/sveltejs/sapper/issues/893
- https://github.com/sveltejs/sapper/pull/869

hoping to submit a repro of the issue to sapper soon