this fixes things a bit
i didn't realize i was using the same function to make the entry href
as i was for the order param href. uh. i understand this better than i
have the time to write about right now. this still breaks if:

- i load (see: refresh) the page with some sort query param already
  set
- i try to click the sort link for that query param

it's because right now sort is a reactive variable. but... when we
first load the page we're getting the server render version, where
query is default. so we need to reset that on the client. but i don't
wanna do that right now!!!