add test for sorting directory (#30)
* add test for sorting directory

* this fixes things a bit

i didn't realize i was using the same function to make the entry href
as i was for the order param href. uh. i understand this better than i
have the time to write about right now. this still breaks if:

- i load (see: refresh) the page with some sort query param already
  set
- i try to click the sort link for that query param

it's because right now sort is a reactive variable. but... when we
first load the page we're getting the server render version, where
query is default. so we need to reset that on the client. but i don't
wanna do that right now!!!

* Delete service-worker.ts

* Revert "Delete service-worker.ts"

This reverts commit 310a22726bb13ccff1af24eac68d4618e9ce94f5.

this didn't seem to help ci

* try using the chrome browser

* back to electron

* try not forcing reload

* update some deps

* update sveltekit version on directoyr page

* Update steps.js

* Update Directory.feature